"Refactor ChatGLM to use Double instead of Float for the temperature parameter in the init method and ChatGLMAPIWrapper struct."

+++ Sources/LangChain/llms/ChatGLM.swift
@@ -14,7 +14,7 @@ public struct ChatGLM: LLM{
     
     let api: ChatGLMAPIWrapper
     
-    public init(model: ChatGLMModel = ChatGLMModel.chatglm_std, temperature: Float = 0.0) {
+    public init(model: ChatGLMModel = ChatGLMModel.chatglm_std, temperature: Double = 0.0) {
         api = ChatGLMAPIWrapper(model: model, temperature: temperature)
     }
